# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Queue API
  description: |
    A distributed task queue system using Redis Streams for reliable message delivery.

    Features:
    - Priority-based task processing (critical, high, normal, low)
    - Scheduled tasks for delayed execution
    - Automatic retries with exponential backoff
    - Dead letter queue for failed tasks
    - Real-time WebSocket events
    - Worker management (pause/resume)
    - Rate limiting and backpressure
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Get task by ID
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Tasks
      summary: Cancel a task
      description: Cancels a pending, scheduled, or retrying task. Cannot cancel running, completed, or failed tasks.
      operationId: cancelTask
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Task cannot be cancelled in current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/workers/{workerId}:
    get:
      tags:
        - Admin
      summary: Get worker details
      operationId: getWorker
      parameters:
        - $ref: '#/components/parameters/workerId'
      responses:
        '200':
          description: Worker details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerInfo'
        '404':
          description: Worker not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/workers/{workerId}/pause:
    post:
      tags:
        - Admin
      summary: Pause a worker
      description: Temporarily stops the worker from processing new tasks
      operationId: pauseWorker
      parameters:
        - $ref: '#/components/parameters/workerId'
      responses:
        '200':
          description: Worker paused
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: worker paused
                  worker_id:
                    type: string
        '404':
          description: Worker not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/workers/{workerId}/resume:
    post:
      tags:
        - Admin
      summary: Resume a worker
      description: Resumes task processing for a paused worker
      operationId: resumeWorker
      parameters:
        - $ref: '#/components/parameters/workerId'
      responses:
        '200':
          description: Worker resumed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: worker resumed
                  worker_id:
                    type: string
        '404':
          description: Worker not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/queues/{priority}:
    delete:
      tags:
        - Admin
      summary: Purge a queue
      description: Removes all messages from the specified priority queue
      operationId: purgeQueue
      parameters:
        - $ref: '#/components/parameters/priority'
      responses:
        '200':
          description: Queue purged
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: queue purged
                  priority:
                    type: string
        '400':
          description: Invalid priority
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/tasks/{taskId}/retry:
    post:
      tags:
        - Admin
      summary: Retry a failed task
      description: Manually retries a failed or dead_letter task
      operationId: retryTask
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task re-queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: task re-queued
                  task_id:
                    type: string
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Task cannot be retried in current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/dlq/retry:
    post:
      tags:
        - Admin
      summary: Retry DLQ tasks
      description: Re-queues one or all tasks from the dead letter queue
      operationId: retryDLQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetryDLQRequest'
      responses:
        '200':
          description: Tasks re-queued
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/RetryDLQSingleResponse'
                  - $ref: '#/components/schemas/RetryDLQAllResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found in DLQ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    taskId:
      name: taskId
      in: path
      required: true
      description: Task identifier
      schema:
        type: string
        format: uuid
    workerId:
      name: workerId
      in: path
      required: true
      description: Worker identifier
      schema:
        type: string
    priority:
      name: priority
      in: path
      required: true
      description: Queue priority level
      schema:
        type: string
        enum: [critical, high, normal, low]
  schemas:
...
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        payload:
          type: object
          additionalProperties: true
        priority:
          type: string
          enum: [low, normal, high, critical]
        state:
          type: string
          enum: [pending, scheduled, running, completed, failed, retrying, cancelled, dead_letter]
        attempts:
          type: integer
        max_retries:
          type: integer
        error:
          type: string
        result:
          type: object
          additionalProperties: true
        worker_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        scheduled_at:
          type: string
          format: date-time
          description: When the task is scheduled to run (for delayed/scheduled tasks and retrying tasks).
        next_retry_at:
          type: string
          format: date-time
          description: When the next retry attempt is scheduled. Only present when state is `retrying`.
        metadata:
          type: object
          additionalProperties:
            type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: HTTP status text
          example: Bad Request
        message:
          type: string
          description: Detailed error message
          example: task type is required
...
    WorkerInfo:
      type: object
      properties:
        id:
          type: string
        state:
          type: string
          enum: [idle, busy, paused, shutting_down]
        started_at:
          type: string
          format: date-time
        last_heartbeat:
          type: string
          format: date-time
        active_tasks:
          type: integer
        concurrency:
          type: integer
        version:
          type: string
...
    RetryDLQRequest:
      type: object
      properties:
        task_id:
          type: string
          description: Specific task ID to retry
        retry_all:
          type: boolean
          description: Set to true to retry all DLQ tasks
        message_id:
          type: string
          description: Optional stream message ID
    RetryDLQSingleResponse:
      type: object
      properties:
        message:
          type: string
          example: task re-queued
        task_id:
          type: string
    RetryDLQAllResponse:
      type: object
      properties:
        message:
          type: string
          example: tasks re-queued
        retried_count:
          type: integer
...